Task1/requirements.txt
datasets==3.2.0
evaluate==0.4.3
matplotlib==3.10.0
numpy==2.2.1
pandas==2.2.3
python-dotenv==1.0.1
torch==2.5.1+cu118
transformers==4.48.0
accelerate==1.2.1